完成界面测试逻辑
